+++ bee/interpreter.py
@@ -1,5 +1,6 @@
 from bee.memory_manager import MemoryManager
+from bee.intermidiate_language import IL
 
 
-def translate(mm: MemoryManager, il: str) -> str:
-    return il
+def translate(mm: MemoryManager, il: IL) -> str:
+    pass